Clean: first clean of alignak_test module
